Refactor to decouple consensus modes
- Functionality that handles Shelley-based eras in a multi-era consensus
  is moved to ouroboros-consensus-shelley so that it can be shared
  between mainnet consensus (in ouroboros-consensus-cardano) and other
  consensus modes comprising multiple Shelley eras.
- The `Protocol` GADT that enumerates consensus modes and provides a
  wealth of information for clients is eliminated. It is an
  implementation detail only used by `cardano-node`. This way,
  `ouroboros-consensus-cardano` does not need to know about other
  consensus modes.
