+++ ouroboros-consensus-cardano-test/src/Test/ThreadNet/TxGen/Cardano.hs
@@ -49,8 +49,8 @@ import qualified Shelley.Spec.Ledger.UTxO as SL (makeWitnessVKey)
 
 import           Ouroboros.Consensus.Shelley.Ledger (GenTx, ShelleyBlock,
                      mkShelleyTx)
-import           Ouroboros.Consensus.Shelley.Ledger.Ledger
-                     (tickedShelleyLedgerState)
+import           Ouroboros.Consensus.Shelley.Ledger.Ledger (Ticked,
+                     tickedShelleyLedgerState)
 
 import           Ouroboros.Consensus.Cardano
 import           Ouroboros.Consensus.Cardano.Block (CardanoEras, GenTx (..),
