--- ouroboros-consensus-cardano/src/Ouroboros/Consensus/Cardano/ShelleyHFC.hs
@@ -1,74 +1,0 @@
-{-# LANGUAGE DataKinds           #-}
-{-# LANGUAGE FlexibleInstances   #-}
-{-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE TypeApplications    #-}
-{-# OPTIONS_GHC -Wno-orphans #-}
-module Ouroboros.Consensus.Cardano.ShelleyHFC (
-    ShelleyBlockHFC
-  ) where
-
-import qualified Data.Map.Strict as Map
-import           Data.SOP.Strict
-
-import           Ouroboros.Consensus.Config
-import           Ouroboros.Consensus.Node.NetworkProtocolVersion
-
-import           Ouroboros.Consensus.HardFork.Combinator
-import           Ouroboros.Consensus.HardFork.Combinator.Serialisation.Common
-
-import           Ouroboros.Consensus.Shelley.Ledger
-import           Ouroboros.Consensus.Shelley.Node ()
-import           Ouroboros.Consensus.Shelley.Protocol
-
-import           Ouroboros.Consensus.Cardano.CanHardFork
-import           Ouroboros.Consensus.Cardano.Node ()
-
-{-------------------------------------------------------------------------------
-  Synonym for convenience
--------------------------------------------------------------------------------}
-
--- | Shelley as the single era in the hard fork combinator
-type ShelleyBlockHFC era = HardForkBlock '[ShelleyBlock era]
-
-{-------------------------------------------------------------------------------
-  NoHardForks instance
--------------------------------------------------------------------------------}
-
-instance ShelleyBasedEra era => NoHardForks (ShelleyBlock era) where
-  getEraParams =
-        shelleyEraParamsNeverHardForks
-      . shelleyLedgerGenesis
-      . configLedger
-  toPartialConsensusConfig _  = tpraosParams
-  toPartialLedgerConfig _ cfg = ShelleyPartialLedgerConfig {
-        shelleyLedgerConfig    = cfg
-      , shelleyTriggerHardFork = TriggerHardForkNever
-      }
-
-{-------------------------------------------------------------------------------
-  SupportedNetworkProtocolVersion instance
--------------------------------------------------------------------------------}
-
--- | Forward to the ShelleyBlock instance. Only supports
--- 'HardForkNodeToNodeDisabled', which is compatible with nodes running with
--- 'ShelleyBlock'.
-instance ShelleyBasedEra era
-      => SupportedNetworkProtocolVersion (ShelleyBlockHFC era) where
-  supportedNodeToNodeVersions _ =
-      Map.map HardForkNodeToNodeDisabled $
-      supportedNodeToNodeVersions (Proxy @(ShelleyBlock era))
-
-  supportedNodeToClientVersions _ =
-      Map.map HardForkNodeToClientDisabled $
-      supportedNodeToClientVersions (Proxy @(ShelleyBlock era))
-
-  latestReleasedNodeVersion = latestReleasedNodeVersionDefault
-
-{-------------------------------------------------------------------------------
-  SerialiseHFC instance
--------------------------------------------------------------------------------}
-
--- | Use the default implementations. This means the serialisation of blocks
--- includes an era wrapper. Each block should do this from the start to be
--- prepared for future hard forks without having to do any bit twiddling.
-instance ShelleyBasedEra era => SerialiseHFC '[ShelleyBlock era]
